pyproject.toml
[tool.poetry]
name = "effectivemobiletesttusk"
version = "0.1.0"
description = ""
authors = ["Andreev <151520440+DeadGEEK990@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
anyio = "4.9.0"
asgiref = "3.8.1"
attrs = "25.3.0"
black = "25.1.0"
build = "1.2.2.post1"
cachecontrol = "0.14.3"
certifi = "2025.6.15"
charset-normalizer = "3.4.2"
cleo = "2.1.0"
click = "8.2.1"
colorama = "0.4.6"
coreapi = "2.3.3"
coreschema = "0.0.4"
crashtest = "0.4.1"
distlib = "0.3.9"
django = "5.2.3"
django-extensions = "4.1"
django-filter = "25.1"
djangorestframework = "3.16.0"
drf-spectacular = "0.28.0"
dulwich = "0.22.8"
fastjsonschema = "2.21.1"
filelock = "3.18.0"
findpython = "0.6.3"
h11 = "0.16.0"
httpcore = "1.0.9"
httpx = "0.28.1"
idna = "3.10"
inflection = "0.5.1"
iniconfig = "2.1.0"
installer = "0.7.0"
itypes = "1.2.0"
jaraco-classes = "3.4.0"
jaraco-context = "6.0.1"
jaraco-functools = "4.2.1"
jinja2 = "3.1.6"
jsonschema = "4.24.0"
jsonschema-specifications = "2025.4.1"
keyring = "25.6.0"
markupsafe = "3.0.2"
more-itertools = "10.7.0"
msgpack = "1.1.1"
mypy-extensions = "1.1.0"
packaging = "25.0"
pathspec = "0.12.1"
pbs-installer = "2025.7.2"
pkginfo = "1.12.1.2"
platformdirs = "4.3.8"
pluggy = "1.6.0"
poetry-core = "2.1.3"
pygments = "2.19.2"
pyproject-hooks = "1.2.0"
pytest = "8.4.1"
pytest-django = "4.11.1"
pywin32-ctypes = "0.2.3"
pyyaml = "6.0.2"
rapidfuzz = "3.13.0"
referencing = "0.36.2"
requests = "2.32.4"
requests-toolbelt = "1.0.0"
rpds-py = "0.26.0"
ruff = "0.12.2"
shellingham = "1.5.4"
sniffio = "1.3.1"
sqlparse = "0.5.3"
tomlkit = "0.13.3"
trove-classifiers = "2025.5.9.12"
typing-extensions = "4.14.1"
tzdata = "2025.2"
uritemplate = "4.2.0"
urllib3 = "2.5.0"
virtualenv = "20.31.2"
zstandard = "0.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
